outsource roihandler in seperate folder and add roi highlighting on hsiplot feature
